Tackling Issue #38. Modularizing gruntfile.
