+++ app/templates/Gruntfile.js
@@ -1,107 +1,14 @@
 'use strict';
 
-
 module.exports = function (grunt) {
+    // Setup access to grunt env variables. Can be accessed via 'grunt.envvars'
+    grunt.envvars = require('./grunt/envvars');
 
-    grunt.initConfig({
-        jshint: {
-            files: ['controllers/**/*.js', 'lib/**/*.js', 'models/**/*.js'],
-            options: {
-                jshintrc: '.jshintrc'
-            }
-        },
-        requirejs: {
-            build: {
-                options: {
-                    baseUrl: 'public/js',
-                    mainConfigFile: 'public/js/config.js',
-                    dir: '.build/js',
-                    optimize: 'uglify',
-                    modules: [{name: 'app'}]
-                }
-            }
-        },
-        less: {
-            build: {
-                options: {
-                    yuicompress: true,
-                    paths: ['public/css']
-                },
-                files: {
-                    '.build/css/app.css': 'public/css/app.less'
-                }
-            }
-        },
-        makara: {
-            files: ['public/templates/**/*.dust'],
-            options: {
-                contentPath: ['locales/**/*.properties']
-            }
-        },
-        dustjs: {
-            build: {
-                files: [
-                    {
-                        expand: true,
-                        cwd: 'tmp/',
-                        src: '**/*.dust',
-                        dest: '.build/templates',
-                        ext: '.js'
-                    }
-                ],
-                options: {
-                    fullname: function (filepath) {
-                        var path = require('path'),
-                            name = path.basename(filepath, '.dust'),
-                            parts = filepath.split(path.sep),
-                            fullname = parts.slice(3, -1).concat(name);
+    // load all grunt plugin tasks matching the `grunt-*` pattern (see: https://github.com/sindresorhus/load-grunt-tasks)
+    require('load-grunt-tasks')(grunt);
 
-                        return fullname.join(path.sep);
-                    }
-                }
-            }
-        },
-        copyto: {
-            build: {
-                files: [
-                    { cwd: 'public', src: ['**/*'], dest: '.build/' }
-                ],
-                options: {
-                    ignore: [
-                        'public/css/**/*',
-                        'public/js/**/*',
-                        'public/templates/**/*'
-                    ]
-                }
-            }
-        },
-        clean: {
-            'tmp': 'tmp',
-            'build': '.build/templates'
-        },
-        mochacli: {
-            src: ['test/*.js'],
-            options: {
-                globals: ['chai'],
-                timeout: 6000,
-                ignoreLeaks: false,
-                ui: 'bdd',
-                reporter: 'spec'
-            }
-        }
-    });
+    // grunt-load-options: will load plugin options and tasks from /grunt/tasks & /grunt/options respective their names (see: https://github.com/chriszarate/grunt-load-options)
 
-    grunt.loadNpmTasks('grunt-contrib-less');
-    <% if (requireJs) { %>grunt.loadNpmTasks('grunt-contrib-requirejs'); <% } %>
-    grunt.loadNpmTasks('grunt-contrib-clean');
-    grunt.loadNpmTasks('grunt-contrib-jshint');
-    grunt.loadNpmTasks('grunt-mocha-cli');
-    grunt.loadNpmTasks('grunt-dustjs');
-    grunt.loadNpmTasks('grunt-copy-to');
+    // Load makara tasks
     grunt.loadTasks('./node_modules/makara/tasks/');
-
-    grunt.registerTask('i18n', ['clean', 'makara', 'dustjs', 'clean:tmp']);
-    grunt.registerTask('build', ['jshint', 'less', <% if (requireJs) { %>'requirejs', <% } %> 'copyto', 'i18n']);
-    grunt.registerTask('test', ['jshint', 'mochacli']);
-
 };
