+++ app/templates/_package.json
@@ -25,6 +25,8 @@
         "mocha": "~1.10.0",
         "supertest": "~0.8.2",
         "grunt": "~0.4.1",
+        "load-grunt-tasks": "~0.2.1",
+        "grunt-load-options": "~0.1.3",
         "grunt-contrib-less": "~0.7.0",
         "grunt-dustjs": "~1.1.1",
         "grunt-contrib-clean": "~0.5.0",
