+++ app/templates/grunt/envvars.js
@@ -1,0 +1,5 @@
+// Environment variables to use through out your tasks.
+// Can be accessed via 'grunt.envvars.prop'
+module.exports = {
+    "prop" : "prop"
+};
