+++ app/templates/grunt/options/clean.js
@@ -1,0 +1,4 @@
+module.exports = {
+    'tmp': 'tmp',
+    'build': '.build/templates'
+};
