+++ app/templates/grunt/options/copyto.js
@@ -1,0 +1,14 @@
+module.exports = {
+    build: {
+        files: [
+            { cwd: 'public', src: ['**/*'], dest: '.build/' }
+        ],
+        options: {
+            ignore: [
+                'public/css/**/*',
+                'public/js/**/*',
+                'public/templates/**/*'
+            ]
+        }
+    }
+};
