+++ app/templates/grunt/options/jshint.js
@@ -1,0 +1,6 @@
+module.exports = {
+    files: ['controllers/**/*.js', 'lib/**/*.js', 'models/**/*.js'],
+    options: {
+        jshintrc: '.jshintrc'
+    }
+};
