+++ app/templates/grunt/options/less.js
@@ -1,0 +1,11 @@
+module.exports = {
+    build: {
+        options: {
+            yuicompress: true,
+            paths: ['public/css']
+        },
+        files: {
+            '.build/css/app.css': 'public/css/app.less'
+        }
+    }
+};
