+++ app/templates/grunt/options/makara.js
@@ -1,0 +1,6 @@
+module.exports = {
+    files: ['public/templates/**/*.dust'],
+    options: {
+        contentPath: ['locales/**/*.properties']
+    }
+};
