+++ app/templates/grunt/options/mochacli.js
@@ -1,0 +1,10 @@
+module.exports = {
+    src: ['test/*.js'],
+    options: {
+        globals: ['chai'],
+        timeout: 6000,
+        ignoreLeaks: false,
+        ui: 'bdd',
+        reporter: 'spec'
+    }
+};
