+++ app/templates/grunt/options/requirejs.js
@@ -1,0 +1,11 @@
+module.exports = {
+    build: {
+        options: {
+            baseUrl: 'public/js',
+            mainConfigFile: 'public/js/config.js',
+            dir: '.build/js',
+            optimize: 'uglify',
+            modules: [{name: 'app'}]
+        }
+    }
+};
