+++ app/templates/grunt/tasks/build.js
@@ -1,0 +1,3 @@
+module.exports = function(grunt) {
+    grunt.registerTask('build', ['jshint', 'less', 'requirejs', 'copyto', 'i18n']);
+};
