+++ app/templates/grunt/tasks/i18n.js
@@ -1,0 +1,3 @@
+module.exports = function(grunt) {
+    grunt.registerTask('i18n', ['clean', 'makara', 'dustjs', 'clean:tmp']);
+};
