+++ app/templates/grunt/tasks/test.js
@@ -1,0 +1,3 @@
+module.exports = function(grunt) {
+    grunt.registerTask('test', ['jshint', 'mochacli']);
+};
